retrieveEmojiMap throws error instead of returning null, used URL streams for data retrieving, changed some formatting
retrieveEmojiMap now throws a ClassCastException instead of returning null and printing an error when the object read is not a valid type
Changed some print statements, nothing major
Letting Jsoup connect to the URL caused some of the HTML received to be cut and thus affected some of the emoji data retrieved, parsed, and saved. Now URL#openStream() is utilized and a FileOutputStream is used to store the HTML received into a temporary cache file. Jsoup is then utilized to read from that cache file to parse the HTML received.
Imported java.net package